Blockeduc/geth/python/requirements.txt
web3==6.11.0
py-solc-x==2.0.2
eth-account==0.10.0
python-dotenv==1.0.0